+++ NEWS.md
@@ -1,3 +1,5 @@
+# huxtable (development version)
+
 
 # huxtable 5.4.0